v.0.1.2.
1) Изменен файл конфигурации
2) В файле конфигурации теперь можно указать настройки цветов для показаний с датчиков
3) Исправлены проблемы с путями. Теперь программа работает по пути C:\Program Files\HovyMonitorBar
4) Доработан установщик, он перемещает файлы автоматически если они расположены не в рабочей директории

+++ HovyMonitor.DeskBar.Win/install_script.bat
@@ -10,7 +10,20 @@ if %errorLevel% == 0 (
 ) else (
     echo Please run this script with administrator permissions.
 	pause
-    goto EXIT
+    exit /b 0
+)
+
+IF "%CD%" == "C:\Program Files\HovyMonitorBar" (
+    echo Directory is correct
+) else (
+    echo You should unpack files right into C:\Program Files\HovyMonitorBar directory.
+    echo "Hit any key to move all files automatically or exit and do it manually"
+    pause
+    mkdir "C:\Program Files\HovyMonitorBar"
+    MOVE "%CD%\*" "C:\Program Files\HovyMonitorBar"
+    call "C:\Program Files\HovyMonitorBar\install_script.bat"
+    exit /b 0
+    pause
 )
 
 taskkill /im explorer.exe /f
